Fixed hidden merge conflict and cyclic import
