+++ director/director_api.go
@@ -20,28 +20,17 @@ package director
 
 import (
 	"context"
-	"encoding/json"
-	"fmt"
-	"net/http"
-	"net/url"
 	"strings"
 	"sync"
 	"time"
 
 	"github.com/jellydator/ttlcache/v3"
-	"github.com/lestrrat-go/httprc"
 	"github.com/lestrrat-go/jwx/v2/jwa"
-	"github.com/lestrrat-go/jwx/v2/jwk"
 	"github.com/lestrrat-go/jwx/v2/jwt"
 	"github.com/pelicanplatform/pelican/config"
 	"github.com/pelicanplatform/pelican/param"
 	"github.com/pkg/errors"
 	log "github.com/sirupsen/logrus"
-)
-
-var (
-	directorJWK      *jwk.Cache
-	directorMetadata *httprc.Cache
 )
 
 // List all namespaces from origins registered at the director
@@ -73,67 +62,6 @@ func ListServerAds(serverTypes []ServerType) []ServerAd {
 		}
 	}
 	return ads
-}
-
-// Return director's public JWK for token verification. This function can be called
-// on any server (director/origin/registry) as long as the Federation_DirectorUrl is set
-//
-// The director's metadata discovery endpoint and JWKS endpoint are cached
-func LoadDirectorPublicKey() (jwk.Key, error) {
-	directorDiscoveryUrlStr := param.Federation_DirectorUrl.GetString()
-	if len(directorDiscoveryUrlStr) == 0 {
-		return nil, errors.Errorf("Director URL is unset; Can't load director's public key")
-	}
-	log.Debugln("Director's discovery URL:", directorDiscoveryUrlStr)
-	directorDiscoveryUrl, err := url.Parse(directorDiscoveryUrlStr)
-	if err != nil {
-		return nil, errors.Wrap(err, fmt.Sprintln("Invalid director URL:", directorDiscoveryUrlStr))
-	}
-	directorDiscoveryUrl.Scheme = "https"
-	directorDiscoveryUrl.Path = directorDiscoveryUrl.Path + directorDiscoveryPath
-
-	directorMetadataCtx := context.Background()
-	if directorMetadata == nil {
-		client := &http.Client{Transport: config.GetTransport()}
-		directorMetadata = httprc.NewCache(directorMetadataCtx)
-		if err := directorMetadata.Register(directorDiscoveryUrl.String(), httprc.WithMinRefreshInterval(15*time.Minute), httprc.WithHTTPClient(client)); err != nil {
-			return nil, errors.Wrap(err, "Failed to register cache for director's metadata")
-		}
-	}
-
-	payload, err := directorMetadata.Get(directorMetadataCtx, directorDiscoveryUrl.String())
-	if err != nil {
-		return nil, errors.Wrap(err, "Failed to get director's metadata")
-	}
-
-	metadata := DiscoveryResponse{}
-
-	err = json.Unmarshal(payload.([]byte), &metadata)
-	if err != nil {
-		return nil, errors.Wrap(err, fmt.Sprintln("Failure when parsing director metadata at: ", directorDiscoveryUrl))
-	}
-
-	jwksUri := metadata.JwksUri
-
-	directorJwkCtx := context.Background()
-	if directorJWK == nil {
-		client := &http.Client{Transport: config.GetTransport()}
-		directorJWK = jwk.NewCache(directorJwkCtx)
-		if err := directorJWK.Register(jwksUri, jwk.WithRefreshInterval(15*time.Minute), jwk.WithHTTPClient(client)); err != nil {
-			return nil, errors.Wrap(err, "Failed to register cache for director's public JWKS")
-		}
-	}
-
-	jwks, err := directorJWK.Get(directorJwkCtx, jwksUri)
-	if err != nil {
-		return nil, errors.Wrap(err, "Failed to get director's public JWKS")
-	}
-	key, ok := jwks.Key(0)
-	if !ok {
-		return nil, errors.Wrap(err, fmt.Sprintln("Failure when getting director's first public key: ", jwksUri))
-	}
-
-	return key, nil
 }
 
 // Create a token for director's Prometheus instance to access
